African Hub Commit (Stage 1b)
African Hub International is a Pan-African business and social
enterprise, with focus on entrepreneurship, culture, creativity and
lifestyle.
In business, our focus is to foster the growth of Africa's emerging
entrepreneurs and SME Sector and we achieve this by providing a platform
and business support services for start-ups, micro, small and medium
enterprises.